# Add benchmark tests to node datastore