--- a/central/node/datastore/singleton.go
+++ b/central/node/datastore/singleton.go
@@ -4,7 +4,7 @@ import (
 	"github.com/stackrox/rox/central/globaldb"
 	"github.com/stackrox/rox/central/globaldb/dackbox"
 	"github.com/stackrox/rox/central/globalindex"
-	"github.com/stackrox/rox/central/node/datastore/dackbox/datastore/keyfence"
+	"github.com/stackrox/rox/central/node/datastore/keyfence"
 	"github.com/stackrox/rox/central/node/datastore/search"
 	"github.com/stackrox/rox/central/node/datastore/store/postgres"
 	"github.com/stackrox/rox/central/ranking"