## Description

Package refactor (first commit) and benchmark tests (second commit).
 
Benchmark results:
```
After:
BenchmarkNodes/upsertNodeWithOldScan
BenchmarkNodes/upsertNodeWithOldScan-8         	       9	 112228287 ns/op
BenchmarkNodes/upsertNodeWithNewScan
BenchmarkNodes/upsertNodeWithNewScan-8         	       8	 312697667 ns/op
BenchmarkNodes/searchAll
BenchmarkNodes/searchAll-8                     	1000000000	         0.0001743 ns/op
BenchmarkNodes/searchForCluster
BenchmarkNodes/searchForCluster-8              	1000000000	         0.0006685 ns/op
BenchmarkNodes/deleteForClusters
BenchmarkNodes/deleteForClusters-8             	   13730	     80150 ns/op

Before:
BenchmarkNodes/upsertNodeWithOldScan
BenchmarkNodes/upsertNodeWithOldScan-8         	       8	 160853915 ns/op
BenchmarkNodes/upsertNodeWithNewScan
BenchmarkNodes/upsertNodeWithNewScan-8         	       7	 166447309 ns/op
BenchmarkNodes/searchAllStoresGetStoresAndListNodes
BenchmarkNodes/searchAllStoresGetStoresAndListNodes-8         	1000000000	         0.0001730 ns/op
BenchmarkNodes/searchForClusterWithListNodes
BenchmarkNodes/searchForClusterWithListNodes-8                	1000000000	         0.0007372 ns/op
BenchmarkNodes/deleteForClustersWithRemoveClusterNodeStore
BenchmarkNodes/deleteForClusters-8             	    7135	    223740 ns/op

```

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps
- [ ] Documented user facing changes (create PR based on [openshift/openshift-docs](https://github.com/openshift/openshift-docs) and merge into [rhacs-docs](https://github.com/openshift/openshift-docs/tree/rhacs-docs))

If any of these don't apply, please comment below.

## Testing Performed
CI