@md2119 2023-01-20
@RTann updated the benchmark results. The node pruning for the cluster improved dramatically. Post cluster remove op was starting point for the refactor effort.